fix classes
