+++ ParkingLock-iOS/Classes/BLEManager.swift
