+++ ParkingLock-iOS/Classes/Ext+Data.swift
@@ -1,0 +1,24 @@
+//
+//  Ext+Data.swift
+//  ParkingLock-iOS
+//
+//  Created by Rilwanul Huda on 01/04/23.
+//
+
+import Foundation
+
+extension Data {
+    func hexEncodedString() -> String {
+        let hexDigits = Array("0123456789abcdef".utf16)
+        var hexChars = [UTF16.CodeUnit]()
+        hexChars.reserveCapacity(count * 2)
+        
+        for byte in self {
+            let (index1, index2) = Int(byte).quotientAndRemainder(dividingBy: 16)
+            hexChars.append(hexDigits[index1])
+            hexChars.append(hexDigits[index2])
+        }
+        
+        return String(utf16CodeUnits: hexChars, count: hexChars.count)
+    }
+}
